🔧 Chore: 패키지 이름 수정 recruite -> recruit
