+++ src/main/java/gaji/service/domain/User.java
@@ -8,7 +8,7 @@ import gaji.service.domain.enums.SocialType;
 import gaji.service.domain.enums.UserActive;
 import gaji.service.domain.message.Message;
 import gaji.service.domain.post.entity.*;
-import gaji.service.domain.recruite.*;
+import gaji.service.domain.recruit.*;
 import gaji.service.domain.room.entity.Event;
 import gaji.service.domain.room.entity.VoiceChatUser;
 import gaji.service.domain.roomPost.*;
