+++ src/main/java/gaji/service/domain/recruit/RecruitPostBookmark.java
@@ -1,5 +1,6 @@
-package gaji.service.domain.recruite;
+package gaji.service.domain.recruit;
 
+import gaji.service.domain.User;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -8,16 +9,18 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class SelectCategory {
+public class RecruitPostBookmark {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @ManyToOne(fetch =  FetchType.LAZY)
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id")
+    private User user;
+
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "post_id")
     private RecruitPost recruitPost;
 
-    @ManyToOne(fetch =  FetchType.LAZY)
-    @JoinColumn(name = "category_id")
-    private Category category;
+
 }
