+++ src/main/java/gaji/service/domain/recruit/RecruitPostLikes.java
@@ -1,4 +1,4 @@
-package gaji.service.domain.recruite;
+package gaji.service.domain.recruit;
 
 import gaji.service.domain.User;
 import jakarta.persistence.*;
@@ -6,13 +6,11 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.time.LocalDateTime;
-
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class SearchKeyword {
 
+public class RecruitPostLikes {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
@@ -21,8 +19,8 @@ public class SearchKeyword {
     @JoinColumn(name = "user_id")
     private User user;
 
-    private String keyword;
-    private LocalDateTime searchDate;
-
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "post_id")
+    private RecruitPost recruitPost;
 
 }
