+++ src/main/java/gaji/service/domain/recruit/SearchKeyword.java
@@ -1,4 +1,4 @@
-package gaji.service.domain.recruite;
+package gaji.service.domain.recruit;
 
 import gaji.service.domain.User;
 import jakarta.persistence.*;
@@ -6,11 +6,13 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.time.LocalDateTime;
+
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class SearchKeyword {
 
-public class RecruitPostLikes {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
@@ -19,8 +21,8 @@ public class RecruitPostLikes {
     @JoinColumn(name = "user_id")
     private User user;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "post_id")
-    private RecruitPost recruitPost;
+    private String keyword;
+    private LocalDateTime searchDate;
+
 
 }
